LayoutTests/http/tests/site-isolation/inspector/dom/query-commands-frame-target-expected.txt

Tests FrameDOMAgent read-only queries: querySelector, querySelectorAll, getOuterHTML, and the search flow.



== Running test suite: SiteIsolation.DOM.FrameDOMAgent.Queries
-- Running test case: Setup
PASS: Should find a cross-origin frame target.
PASS: Frame target should have a document.
PASS: Should find #container.

-- Running test case: SiteIsolation.DOM.Queries.QuerySelector
PASS: Should get a nodeId for '#heading'.
PASS: Frontend should resolve the nodeId to a DOMNode.
PASS: Resolved node should be an H1.
PASS: Resolved node id should be 'heading'.

-- Running test case: SiteIsolation.DOM.Queries.QuerySelector.NoMatch
PASS: Should return a falsy nodeId when nothing matches.

-- Running test case: SiteIsolation.DOM.Queries.QuerySelectorAll
PASS: Should return two li.item nodeIds.
PASS: Frontend should resolve each nodeId.
PASS: Resolved node should be LI.
PASS: Frontend should resolve each nodeId.
PASS: Resolved node should be LI.

-- Running test case: SiteIsolation.DOM.Queries.GetOuterHTML
PASS: Should serialize the H1 subtree.

-- Running test case: SiteIsolation.DOM.Queries.GetSupportedEventNames
PASS: Should return an array.
PASS: Event names list should be non-empty.
PASS: Event names should include 'click'.
PASS: Event names should include 'load'.

-- Running test case: SiteIsolation.DOM.Queries.PushNodeByPathToFrontend
PASS: Should resolve the path to a nodeId.
PASS: Resolved node should be #container.

-- Running test case: SiteIsolation.DOM.Queries.Search
PASS: performSearch should return a searchId.
PASS: Search for 'frame-root' should match exactly one node.
PASS: getSearchResults should return resultCount ids.
PASS: Search result should resolve on the frontend.
PASS: Search result should be #container.
PASS: Should produce an exception.
Error: Missing search result for given searchId

-- Running test case: SiteIsolation.DOM.Queries.Search.CaseSensitive
PASS: Case-insensitive search for 'FRAME-ROOT' should match.
PASS: Case-sensitive search for 'FRAME-ROOT' should not match.

-- Running test case: SiteIsolation.DOM.Queries.Search.ScopedToSubtree
PASS: Should find #list for scoping.
PASS: Unscoped search should match the H1.
PASS: Search scoped to #list should not match H1.
PASS: Search scoped to #list should match 'Item'.
